@@ -55,7 +55,7 @@
     "regenerator-runtime": "^0.13.11",
     "sbc-common-components": "3.0.8",
     "uuid": "^9.0.0",
-    "vue": "2.7.14",
+    "vue": "^3.3.4",
     "vue-affix": "^0.5.2",
     "vue-hotjar": "^1.4.0",
     "vue-observe-visibility": "^1.0.0",
@@ -63,7 +63,7 @@
     "vue-the-mask": "^0.11.1",
     "vue2-filters": "^0.14.0",
     "vuelidate": "0.6.2",
-    "vuetify": "^2.6.15",
+    "vuetify": "^3.3.19",
     "vuex": "^3.6.2"
   },
   "devDependencies": {
@@ -75,14 +75,14 @@
     "@types/vuelidate": "^0.7.16",
     "@typescript-eslint/eslint-plugin": "^5.59.2",
     "@typescript-eslint/parser": "^5.59.2",
-    "@vitejs/plugin-vue2": "^2.2.0",
+    "@vitejs/plugin-vue":"^4.3.4",
     "@volar-plugins/vetur": "latest",
     "@vue/eslint-config-standard": "^4.0.0",
     "@vue/eslint-config-typescript": "^9.1.0",
     "@vue/test-utils": "^1.3.5",
     "eslint": "~7.32.0",
-    "eslint-plugin-vue": "^8.7.1",
-    "eslint-plugin-vuetify": "^1.1.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "eslint-plugin-vuetify": "^2.0.5",
     "flush-promises": "^1.0.2",
     "intersection-observer": "^0.12.2",
     "jsdom": "^22.1.0",
