@@ -77,7 +77,7 @@
               class="inherit-checkbox"
               hide-details
               label="Same as Mailing Address"
-              @change="setDeliveryAddressToMailingAddress()"
+              @update:modelValue="setDeliveryAddressToMailingAddress()"
             />
             <template
               v-if="!isSame(mailingAddress, deliveryAddress, ['actions']) || !inheritMailingAddress"
@@ -123,7 +123,7 @@
               class="inherit-checkbox"
               hide-details
               label="Same as Registered Office"
-              @change="setRecordOfficeToRegisteredOffice()"
+              @update:modelValue="setRecordOfficeToRegisteredOffice()"
             />
           </v-col>
         </v-row>
@@ -186,7 +186,7 @@
                   class="inherit-checkbox"
                   hide-details
                   label="Same as Mailing Address"
-                  @change="setRecordDeliveryAddressToMailingAddress()"
+                  @update:modelValue="setRecordDeliveryAddressToMailingAddress()"
                 />
                 <template
                   v-if="!isSame(recMailingAddress, recDeliveryAddress, ['actions']) || !inheritRecMailingAddress"
