@@ -24,7 +24,7 @@
         <v-btn
           id="dialog-exit-button"
           color="primary"
-          text
+          variant="text"
           @click="exit()"
         >
           Exit
@@ -33,7 +33,7 @@
         <v-btn
           id="dialog-retry-button"
           color="primary"
-          text
+          variant="text"
           @click="retry()"
         >
           Retry
